Do not cache InetSocketAddress in DisqueURI #16
The SocketAddress obtained by getResolvedAddress is no longer cached. This is to allow dynamic DNS updates at the time of establishing a connection. This change introduces the possibility to change the Disque connection point details because the connection point details are obtained directly from the DisqueURI argument.

Reference: mp911de/lettuce#144